--- Projeto_3/pessoa.h
@@ -1,24 +1,0 @@
-#ifndef PESSOA_H
-#define PESSOA_H
-
-#include <string>
-#include <iostream>
-
-class Pessoa {
-private:
-    std::string nome;
-    std::string telefone;
-
-public:
-    Pessoa(); // Construtor
-
-    
-    Pessoa(const std::string& nome, const std::string& telefone);
-
-    ~Pessoa();    // Destrutor
-
-    void imprimirNome() const;
-    void imprimirTelefone() const;
-};
-
-#endif
